Extract shared behaviour and views from v2
