+++ app/views/v2/home/local_authorities.html.erb
@@ -53,7 +53,7 @@
       <div class="panel-heading clearfix">
         <div class="row">
           <div class="col-md-3">
-            <%= render 'sort_order_dropdown' %>
+            <%= render 'v2/shared/sort_order_dropdown' %>
           </div>
           <div class="col-md-9">
             <%= select_tag 'filter', options_for_select(LocalAuthority.all.map { |la| [la.name, la.id] }), include_blank: true, id: 'filter-dropdown', class: 'form-control' %>
