+++ app/views/v2/home/services.html.erb
@@ -53,7 +53,7 @@
       <div class="panel-heading">
         <div class="row">
           <div class="col-md-3">
-            <%= render 'sort_order_dropdown' %>
+            <%= render 'v2/shared/sort_order_dropdown' %>
           </div>
           <div class="col-md-9">
             <%= select_tag 'filter', options_for_select(Service.enabled.all.map { |s| [s.label, s.id] }), include_blank: true, id: 'filter-dropdown', class: 'form-control' %>
