+++ app/views/v2/shared/_link_table_row.html.erb
